fixed bug in use memo hook
