+++ app/[lang]/artistry/CompositionTable/MusicTableSelection.tsx
@@ -33,14 +33,35 @@ export default function MusicTableSection({ lang }: Readonly<Props>) {
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
   const borderWithOpacity = hexToRGBA(mainHexPallete.blue[200], 0.4);
   const t = useTranslations('table.name');
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const [search, setSearch] = useState(searchParams.get('search') || '');
+  const [data, setData] = useState<Music[]>([]);
+  const [isLoading, setIsLoading] = useState(false);
 
-  // useEffect(() => {
-  //   setHasMounted(true);
-  // }, []);
+  useEffect(() => {
+    const newParams = new URLSearchParams(searchParams);
+    if (search) {
+      newParams.set('search', search);
+      router.refresh();
+    } else {
+      newParams.delete('search');
+      router.refresh();
+    }
+    router.replace(`?${newParams.toString()}`, { scroll: false });
+  }, [router, search, searchParams, isLoading]);
+  useEffect(() => {
+    const fetchData = async () => {
+      setIsLoading(true);
+      const endpoint = `/api/compositions?lang=${lang}&search=${encodeURIComponent(search)}`;
+      const res = await fetch(endpoint);
+      const json = await res.json();
+      setData(json);
+      setIsLoading(false);
+    };
 
-  // if (!hasMounted || !data || data.length === 0) {
-  //   return null;
-  // }
+    fetchData();
+  }, [search, lang]);
   const columns = [
     { id: 'expander', header: '', cell: () => null },
     {
@@ -85,6 +106,8 @@ export default function MusicTableSection({ lang }: Readonly<Props>) {
   ];
   return (
     <EnhancedTable
+      data={data}
+      loading={isLoading}
       columns={columns}
       groupByKey="opus"
       columnFilters={columnFilters}
@@ -100,7 +123,7 @@ export default function MusicTableSection({ lang }: Readonly<Props>) {
       }}
       itemsPerPage={10}
       tableName={t('composition')}
-      // MusicSearch={<MusicSearch search={search} setSearch={setSearch} />}
+      MusicSearch={<MusicSearch search={search} setSearch={setSearch} />}
     />
   );
 }
